+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## [UNRELEASED]
 
 
+## [1.9.0] - 2018-2-5
+
+- Add `Domain#validateInput` and `Domain#validateOutput` methods
+- Remove ambiguous `protocol#validate` method
+
+## [1.8.0] - 2018-2-5
+
+- Add `Def.isPII` method and `PII` label for Types
+
 ## [1.7.1] - 2018-1-11
 
 - Add "Allowed values:" to prop enum docs